# Conflicts:
#	src/client/app/data/initStore.ts
#	src/client/app/demo_view/chat/chat.tpl
#	src/client/app/demo_view/chat/chat.ts
#	src/client/app/widget/messageItem/messageItem.tpl
#	src/client/app/widget/messageItem/messageItem.ts
#	src/server/data/rpc/basic.r.ts

+++ src/server/data/rpc/message.rs
@@ -5,19 +5,19 @@ struct AnnounceSend {
     gid: u32,//组id
     mtype: MSG_TYPE,
     msg: String,//内容
-    time: u32,//时间
+    time: usize,//时间
 }
 
 struct UserSend {
     rid: u32,//接受者id
     mtype: MSG_TYPE,
     msg: String,//内容
-    time: u32,//时间
+    time: usize,//时间
 }
 
 struct GroupSend {
     gid: u32,//组id
     mtype: MSG_TYPE,
     msg: String,//内容,可能是url
-    time: u32,//时间
+    time: usize,//时间
 }
